requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
aiomysql
